// Front-End/src/app/interfaces/registro-response.ts
// Generated by https://quicktype.io

export interface RegistroResponse {
    ok:         boolean;
    msg:        string;
    usuarioDB:  UsuarioDB;
}

export interface UsuarioDB {
    id_usuario:                    number;
    nombres:                       string;
    primer_apellido:               string;
    segundo_apellido:              string;
    fecha_nacimiento:              string;
    edad:                          string;
    correo_electronico:            string;
    contrasena:                    string;
    curp:                          string;
    uuid:                          string;
    estado:                        number;
    perfil:                        number;
    genero:                        null;
    telefono_celular:              null;
    telefono_casa:                 null;
    demarcacion_territorial:       number;
    entidad_federativa:            number;
    como_te_enteraste:             null;
    id_comite:                     null;
    folio:                         null;
    fecha_selecciona_comite:       null;
    fecha_alta:                    string;
    fecha_baja:                    null;
    fecha_activa_cuenta:           null;
    actualiza_contrasena:          null;
    fecha_actualiza_contrasena:    null;
    token_mail:                    null;
    fecha_token:                   null;
    fecha_complementa_informacion: null;
}